requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==7.1.2
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
frozendict==2.4.6
html5lib==1.1
idna==3.10
joblib==1.4.2
jugaad-data==0.26
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
multitasking==0.0.11
numpy==2.1.3
packaging==24.1
pandas==2.2.3
peewee==3.17.7
pillow==11.0.0
platformdirs==4.3.6
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sklearn-model==0.0.6
soupsieve==2.6
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
yfinance==0.2.48
